Add equipment_model_state_hourly_earnings model

+++ code/AIKO_TestProject/AIKO_TestProject/Models/EquipmentPositionHistory.cs
@@ -10,8 +10,8 @@ namespace AIKO_TestProject.Models
     public class EquipmentPositionHistory
     {
         public Guid equipment_id { get; set; }
-        public int lat { get; set; }
-        public int lon { get; set; }
+        public Single lat { get; set; }
+        public Single lon { get; set; }
         public DateTime date { get; set; }
 
     }
